refactor: useQuiz, useQuizResult hook 경로 변경
